fetch quotes of loginUser
