+++ src/contexts/Utils/interface.ts
@@ -2,3 +2,7 @@ export type UtilsContextType = {
   isMobile: boolean;
   handleResize: () => void;
 };
+
+export interface Props {
+  children: React.ReactNode;
+}
