+++ src/pages/Header/EditBtn.tsx
@@ -1,6 +1,6 @@
 import { IconButton, Tooltip } from "@mui/material";
 import { AiOutlineEdit } from "react-icons/ai";
-import { useModalContext } from "../../contexts/ModalContext";
+import { useModalContext } from "../../contexts/Modal/ModalContext";
 import { BiEditAlt } from "react-icons/bi";
 
 const EditBtn = () => {
