+++ src/pages/Modal/GoPrev.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useModalContext } from "../../contexts/ModalContext";
+import { useModalContext } from "../../contexts/Modal/ModalContext";
 import { IoIosArrowBack } from "react-icons/io";
 
 const GoPrev = () => {
@@ -8,7 +8,7 @@ const GoPrev = () => {
   return (
     <div
       className="cursor-pointer flex items-center gap-2 p-1 hover:opacity-30"
-      onClick={() => handleCurrPageNum('prev')}
+      onClick={() => handleCurrPageNum("prev")}
     >
       <IoIosArrowBack />
       <span>Back</span>
