+++ src/pages/Modal/InputsPage2/CdCreate/EventDateInput.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useCountdownContext } from "../../../../contexts/CountdownContext";
+import { useCountdownContext } from "../../../../contexts/Countdown/CountdownContext";
 import { DesktopDatePicker } from "@mui/x-date-pickers/DesktopDatePicker";
 
 import TextField from "@mui/material/TextField";
