+++ src/pages/Modal/InputsPage2/CdCreate/EventTitleInput.tsx
@@ -1,5 +1,5 @@
 import TextField from "@mui/material/TextField";
-import { useCountdownContext } from "../../../../contexts/CountdownContext";
+import { useCountdownContext } from "../../../../contexts/Countdown/CountdownContext";
 
 const EventTitleInput: React.FC = () => {
   const { handleEventTitleInputText } = useCountdownContext();
