+++ src/pages/Modal/InputsPage2/CdCreate/SaveBtn.tsx
@@ -1,6 +1,6 @@
-import { useAuthContext } from "../../../../contexts/AuthContext";
-import { useCountdownContext } from "../../../../contexts/CountdownContext";
-import { useModalContext } from "../../../../contexts/ModalContext";
+import { useAuthContext } from "../../../../contexts/Auth/AuthContext";
+import { useCountdownContext } from "../../../../contexts/Countdown/CountdownContext";
+import { useModalContext } from "../../../../contexts/Modal/ModalContext";
 
 const SaveBtn = () => {
   const { handleSaveBtnClick } = useCountdownContext();
