+++ src/pages/Modal/InputsPage2/CdEdit/EventDateInput.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { useCountdownContext } from "../../../../contexts/CountdownContext";
+import { useCountdownContext } from "../../../../contexts/Countdown/CountdownContext";
 
 import TextField from "@mui/material/TextField";
 import { DatePicker } from "@mui/x-date-pickers";
