+++ src/pages/Modal/InputsPage2/CdEdit/SaveBtn.tsx
@@ -1,6 +1,6 @@
-import { useAuthContext } from "../../../../contexts/AuthContext";
-import { useCountdownContext } from "../../../../contexts/CountdownContext";
-import { useModalContext } from "../../../../contexts/ModalContext";
+import { useAuthContext } from "../../../../contexts/Auth/AuthContext";
+import { useCountdownContext } from "../../../../contexts/Countdown/CountdownContext";
+import { useModalContext } from "../../../../contexts/Modal/ModalContext";
 
 const SaveBtn = () => {
   const { handleSaveBtnClick } = useCountdownContext();
@@ -9,8 +9,8 @@ const SaveBtn = () => {
   return (
     <button
       onClick={() => {
-        handleSaveBtnClick('edit', loginUser.uid)
-        handleCurrPageNum('done')
+        handleSaveBtnClick("edit", loginUser.uid);
+        handleCurrPageNum("done");
         handleToggleModal();
       }}
       className="w-48 hover:bg-sky-50 hover:text-sky-600 rounded-md p-2 ease-in-out duration-200"
