+++ src/pages/Modal/SelectType1/Cd/CreateBtn.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useModalContext } from "../../../../contexts/ModalContext";
+import { useModalContext } from "../../../../contexts/Modal/ModalContext";
 
 const CreateBtn = () => {
   const { handleSelectTypeAndAct } = useModalContext();
